requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
openpyxl>=3.1.0
google-generativeai>=0.7.0
rapidfuzz>=3.6.0
reportlab>=4.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
xlsxwriter>=3.1.0
numpy>=1.24.0
